+++ CHANGELOG.rst
@@ -2,6 +2,11 @@
 CHANGELOG
 =========
 
+1.10.2dev
+=========
+* bug-fix: Setting health check timeout limit on local mode to 30s
+* bug-fix: Make Hyperparameters in local mode optional.
+
 1.10.1
 ======